@@ -16,7 +16,7 @@
 package com.example.android.didyoufeelit;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.widget.TextView;
 
 /**
